Adds getProfile and createProfile methods to the authAPI.js file.
Adds

$.ajaxSetup({
   xhrFields: {
       withCredentials: true
   }
});

to authAPI.ajax method for sending cookie in the broswer

In chrome console, need to go to network tab and turn on cookie session (righ click on the columns) for the cookie to work.

Now when user log in, user profile is checked. If user profile exists, do nothing; if user has no profile, create a new profile.

Tested it in both browser and postman.

jsm